:sparkles: feat(frontend): group-detail 페이지에서 사용하는 공통 layout 추가 (#320)
