+++ frontend/components/ui/layout/group/GroupDetailFormat.module.scss
@@ -1,0 +1,89 @@
+.container {
+	@screen md {
+		@apply h-full;
+	}
+	@apply h-full text-customDark;
+
+	//컨텐츠
+	.contents_container {
+		@apply flex relative overflow-x-hidden;
+		height: calc(100vh - 80px);
+		overflow-y: auto; /* 스크롤 가능한 컨테이너로 만듭니다. 이걸 해주어야 왼쪽 사이드바 sticky가 먹음 */
+
+		.detail_container {
+			@screen md {
+				flex: 1;
+			}
+			@apply relative w-full z-10;
+
+			.main {
+				@apply h-full;
+				.banner_img_container {
+					@screen md {
+						@apply h-[400px];
+					}
+					@apply relative h-[200px] border-b border-solid border-customDark;
+
+					.banner_edit_btn {
+						@apply absolute flex gap-1 border border-solid border-customDark rounded px-2 py-1 bottom-4 right-4 bg-customOrange font-semibold;
+
+						.btn_text {
+							@apply text-sm cursor-pointer;
+						}
+					}
+				}
+				.main_contents_container {
+					@screen md {
+						@apply mx-auto;
+						width: 768px;
+					}
+					@apply w-full;
+
+					.banner_profile_contaienr {
+						@screen md {
+							@apply mx-0 flex-row gap-0;
+						}
+						@apply flex flex-col gap-4 mt-6 mx-4;
+
+						.banner_profile_right_contaienr {
+							@screen md {
+								@apply ml-auto flex gap-4;
+							}
+							@apply ml-0 block;
+
+							.create_feed_btn {
+								@screen md {
+									@apply block w-[200px];
+								}
+								@apply hidden;
+							}
+
+							.toggle_menu_container {
+								@screen md {
+									@apply w-[200px];
+								}
+
+								@apply flex justify-center items-center relative w-full;
+							}
+						}
+					}
+
+					.tap_menu_container {
+						@screen md {
+							@apply mx-0;
+						}
+						@apply mx-4 z-10 mt-8 mb-6;
+					}
+				}
+			}
+
+			// mobile-create-feed-btn
+			.mobile_create_feed_btn_container {
+				@screen md {
+					@apply hidden;
+				}
+				@apply fixed bottom-6 right-4 border border-solid border-customDark bg-customOrange flex justify-center items-center rounded-full p-4 z-30;
+			}
+		}
+	}
+}
