+++ frontend/components/ui/layout/group/group-detail-format.interface.ts
@@ -1,0 +1,4 @@
+export interface GroupDetailFormatProps {
+	lottieLike: boolean;
+	groupId: string;
+}
